refactor: consolidate and simplify tool execution modules
**Before (confusing structure):**
- `execution_simple.rs` (80 lines) - context types
- `simple_execution.rs` (300+ lines) - execution engine
- Unclear naming: which is which?

**After (clean structure):**
- `execution.rs` (350 lines) - all tool execution in one place
- Clear responsibility: contains context, engine, and result types
- Single module with logical organization

**Changes:**
- Merged `execution_simple.rs` into `simple_execution.rs`
- Renamed `simple_execution.rs` → `execution.rs` (clearer name)
- Updated imports and exports
- Removed redundant module declarations
- All tests still pass

This follows the principle of module organization: related functionality should be together, and module names should clearly indicate their purpose.
