+++ src-tauri/src/claude/types.rs
@@ -73,19 +73,6 @@ impl ContentBlock {
         }
     }
 
-    pub fn tool_result_with_metadata(
-        tool_use_id: impl Into<String>,
-        content: impl Into<String>,
-        is_error: Option<bool>,
-        metadata: crate::claude::tools::ToolResultMetadata,
-    ) -> Self {
-        Self::ToolResult {
-            tool_use_id: tool_use_id.into(),
-            content: content.into(),
-            is_error,
-            metadata: Some(metadata),
-        }
-    }
 
     #[allow(dead_code)]
     pub fn thinking(content: impl Into<String>) -> Self {
